made some changes to the frontend on the technician side. show part price when creating a new device. updated validation of part creation to allow the creation of devices with a price of 0

+++ app/controllers/PartController.php
@@ -138,7 +138,7 @@ class PartController
         try {
             $result = $this->partModel->create(
                 $_POST['name'],
-                intval($_POST['price']),
+                floatval($_POST['price']),
                 intval($_POST['amount']),
                 intval($_POST['part_type'])
             );
